Move assert functions to test_utils.py
